api_yamdb/requirements.txt
requests==2.26.0
Django
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt
django-filter
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv